"Update app routes to include UpdateAd screen and adjust CreateNewAdvertisement params"
This diff includes changes to the app's route configuration to add a new screen for updating ads and adjusting the parameters for the CreateNewAdvertisement screen. The UpdateAd screen has been added to the AppRoutes type and the navigation options. The CreateNewAdvertisement screen has been updated to accept a new propsId parameter instead of an id parameter. The Product screen has also been updated to correct the alt text for the avatar image.